thingspeak_adaptor/requirements.txt
requests
cherrypy
paho-mqtt==1.6.1